+++ src/app/meals/create/recipe-section.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import styles from '@/styles/app/meals/create/recipe-form.module.scss';
 import { MealRecipeSectionWithId } from '@/src/types/meal.types';
 import { InputString } from '@/src/components/common/form/input-string';
 import { useState } from 'react';
@@ -7,6 +8,7 @@ import { useRecipeFormContext } from '@/src/contexts/recipe-form.context';
 import { createNewRecipeStep, updateSectionName, updateSectionStep } from '@/src/helpers/recipe-form.helper';
 import { RecipeSectionStep } from '@/src/app/meals/create/recipe-section-step';
 import { AddButton } from '@/src/components/common/add-button';
+import { RemoveButton } from '@/src/components/common/remove-button';
 
 interface RecipeSectionProps {
     section: MealRecipeSectionWithId;
@@ -32,7 +34,10 @@ export function RecipeSection({ section }: RecipeSectionProps) {
 
     return (
         <li>
-            <InputString label={'Section name'} value={name} setValue={setValue} />
+            <div className={styles['section-name-container']}>
+                <InputString label={'Section name'} value={name} setValue={setValue} />
+                <RemoveButton label={'Delete'} onClick={() => {}} />
+            </div>
             <ol>
                 {section.steps.map(step => <RecipeSectionStep key={step.id} step={step} section={section} />)}
             </ol>